Limit chunk size to 4GiB
Realistically, a chunk size anywhere close to 4GiB is going to be
impractical. It's useful to lock the chunk size to 4GiB for three
reasons:

1. Previously, the on-disk format was storing the chunk size as a usize,
which is serialized using 64-bits, wasting space.
2. Limiting the chunk size to a u32 in the public API prevents the user
from setting the chunk size to something unreasonable.
3. A chunk size larger than 4GiB would be impossible on 32-bit systems.

+++ src/repo/common/chunking.rs
@@ -36,7 +36,7 @@ pub enum Chunking {
     /// content-defined deduplication and will typically result in worse deduplication ratios.
     Fixed {
         /// The size of each chunk in bytes.
-        size: usize,
+        size: u32,
     },
 
     /// Split data using the ZPAQ content-defined chunking algorithm.
@@ -48,7 +48,7 @@ pub enum Chunking {
         ///
         /// For example, a value of `20` will result in an average chunk size of 1MiB
         /// (2^20 = 1048576).
-        bits: usize,
+        bits: u32,
     },
 }
 
@@ -56,8 +56,8 @@ impl Chunking {
     /// Return a chunker for this chunking method.
     pub(super) fn to_chunker(&self) -> Box<dyn ChunkerImpl> {
         match self {
-            Chunking::Fixed { size } => Box::new(FixedChunker::new(*size)),
-            Chunking::Zpaq { bits } => Box::new(ZPAQ::new(*bits)),
+            Chunking::Fixed { size } => Box::new(FixedChunker::new(*size as usize)),
+            Chunking::Zpaq { bits } => Box::new(ZPAQ::new(*bits as usize)),
         }
     }
 }
